Added more Dialogue Line Types
