+++ When the Crow Sings/Assets/Dialogue/TestText.txt
@@ -1,8 +1,11 @@
 ~ my_title
 Character 1: Hello!
 This is text.
+=> END
 
 ~ my_second_title
 Wassupppp
 Yeah
 Character 2: Oh no: not this too!!
+
+=> END
