+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6457bb44aaaa8df4e890a7571fbd35e7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
