+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueBase.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+abstract public class DialogueBase
+{
+    public int nextID;
+}
