+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f10c25d848da0b84597be9f1237e247d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
