+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueChoice.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DialogueChoice : DialogueBase
+{
+    // A potential dialogue response.
+    // ...I feel like there needs to be some sort of container for "all choices that are available at once."
+}
