+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueChoice.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b2daedd6522e220498ea599f6a59ca99
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
