+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueCondition.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a966440d28b0f094c9ea4dd06c8ffa2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
