+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueGoto.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DialogueGoto : DialogueBase
+{
+    public bool isEnd = false;
+
+
+    // if text is "=> END" then "finish the dialogue."
+}
