+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueGoto.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4a6d13f54d4460d489d5ae441f80e33e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
