+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueMutation.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DialogueMutation : DialogueBase
+{
+    // A line that sets a variable, calls a method, or otherwise modifies the game state.
+}
