+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueMutation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a629246b06539fb4fb9367ef869184ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
