+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueTitle.cs
@@ -1,0 +1,7 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DialogueTitle : DialogueBase
+{
+}
