+++ When the Crow Sings/Assets/Scripts/Dialogue/LineTypes/DialogueTitle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b5d9d21d7863a546bd466120516f61c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
